pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agentic-framework"
version = "0.1.0"
description = "Production-ready AI Agentic Framework with composable components"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Selva" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai",
    "pydantic",
    "python-dotenv",
    "chromadb",
    "docker",
    "pinecone-client",
    "redis",
    "aioredis",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation",
    "prometheus-client",
    "aiohttp",
    "asyncio",
    "jsonschema",
    "pyyaml",
    "jinja2",
    "click",
    "rich",
    "yfinance",
    "ddgs",
    "langgraph",
    "langchain-core",
]

[project.scripts]
agentic = "agentic_framework.cli:cli"

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pdoc>=14.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=agentic_framework --cov-report=html --cov-report=term"
testpaths = ["tests"]
asyncio_mode = "auto"
